Warning cleanup
git-svn-id: http://svn.l2jfree.com/l2jfree@5170 6ff00ba9-25cd-4355-987e-099c62bc2628

+++ trunk/l2j-commons/src/main/java/com/l2jfree/tools/algorithm/MultiSort.java
@@ -98,7 +98,7 @@ public class MultiSort {
 		for (Integer currValue : getValues())
 			totalValue += (1.0 / currValue.doubleValue());
 
-		return ((double) getCount() / totalValue);
+		return getCount() / totalValue;
 	}
 
 	public final List<?> getKeys() {
